requirements.txt
discord.py
discord.py[voice]
python-decouple
youtube_dl
ffmpeg
PyNaCl

aiohttp
async-timeout
astroid
cffi
chardet
colorama
cryptography
idna
isort
lazy-object-proxy
mccabe
multidict
pycparser
pylint
pymongo
pyspnego
six
smbprotocol
toml
typing_extensions
wrapt
yarl